The HttpDaemonRspResponse class

Package: R.rsp
Class HttpDaemonRspResponse

Object
~~|
~~+--RspResponse
~~~~~~~|
~~~~~~~+--FileRspResponse
~~~~~~~~~~~~|
~~~~~~~~~~~~+--HttpDaemonRspResponse

Directly known subclasses:

public static class HttpDaemonRspResponse
extends FileRspResponse

An instance of class HttpDaemonRspResponse, which extends the RspResponse class, is a buffer for output (response) sent to an HttpDaemon. It provides a method writeResponse() for writing output and a method flush() for flush the written output to the HTTP daemon.

HttpDaemonRspResponse(httpDaemon=NULL, ...)

Arguments

httpDaemon

An HttpDaemon object.

...

Not used.

Fields and Methods

Methods:

flush	-

writeResponse	-

Methods inherited from FileRspResponse:
flush, getAbsolutePath, getName, getOutput, getPath, writeResponse

Methods inherited from RspResponse:
flush, writeResponse

Methods inherited from Object:
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, names, objectSize, print, save

Details

The purpose of this method is to provide partial writing of HTTP response such that, for instance, a web browser can display parts of an HTML page while the rest is generated. Note that this is only supported by the HTTP v1.1 protocol.

Note: The minimalistic HTTP daemon (written in Tcl) used internally currently only supports HTTP v1.0. In other words, although this class is used already, the output is only flushed at the end.

Author

Henrik Bengtsson

See also