The RspIncludeDirective class

Package: R.rsp
Class RspIncludeDirective

character
~~|
~~+--RspObject
~~~~~~~|
~~~~~~~+--RspConstruct
~~~~~~~~~~~~|
~~~~~~~~~~~~+--RspDirective
~~~~~~~~~~~~~~~~~|
~~~~~~~~~~~~~~~~~+--RspIncludeDirective

Directly known subclasses:

public class RspIncludeDirective
extends RspDirective

An RspIncludeDirective is an RspDirective that causes the RSP parser to include (and parse) an external RSP file.

RspIncludeDirective(value="include", ...)

Arguments

value

A character string.

attributes

A named list, which must contain either a 'file' or a 'text' element.

...

Optional arguments passed to the constructor of RspDirective.

Fields and Methods

Methods:

getContent	Gets the content of the RSP include directive.

getFile	Gets the file attribute.

getVerbatim	Checks if verbatim include should be used or not.

getWrap	Get the wrap length.

Methods inherited from RspDirective:
asRspString, getNameContentDefaultAttributes, requireAttributes

Methods inherited from RspConstruct:
asRspString, getComment, getInclude, getSuffixSpecs

Methods inherited from RspObject:
print

Methods inherited from character:
Ops,nonStructure,vector-method, Ops,structure,vector-method, Ops,vector,nonStructure-method, Ops,vector,structure-method, all.equal, as.Date, as.POSIXlt, as.data.frame, as.raster, coerce,ANY,character-method, coerce,character,SuperClassMethod-method, coerce,character,signature-method, coerce<-,ObjectsWithPackage,character-method, coerce<-,signature,character-method, formula, getDLLRegisteredRoutines, isOpen, toLatex, uses

Author

Henrik Bengtsson