Starts the internal web browser and opens the URL in the default web browser

Starts the internal web browser and opens the URL in the default web browser. From this page you access not only help pages and demos on how to use RSP, but also other package RSP pages.

# S3 method for default
browseRsp(url=paste(urlRoot, path, sep = "/"), urlRoot=sprintf("http://%s:%d",
  host, port), host="127.0.0.1", port=8074L, path="", start=TRUE, stop=FALSE, ...)

Arguments

url
A character string for the URL to be viewed. By default the URL is constructed from the urlRoot and the path parameters.

urlRoot
A character string specifying the URL root. By default the URL is constructed from the host and the port.

host
An optional character string for the host of the URL.

port
An optional integer for the port of the URL.

path
An optional character string for the context path of the URL.

start
If TRUE, the internal R web server is started if not already started, otherwise not.

stop
If TRUE, the internal R web server is stopped, if started.

...
Additional arguments passed to browseURL.

Value

Returns (invisibly) the URL.

Author

Henrik Bengtsson

See also